refactor(studio): compare against exclusive range end; move types to audit.types.ts
- Drop the rangeEnd - 1ms arithmetic in both report queries and compare
  directly against the exclusive boundary (createdAt < rangeEnd,
  deletedAt >= rangeEnd). Postgres timestamps carry microsecond
  precision, so the -1ms inclusive bound misclassified rows in the
  final millisecond of the range.
- Move AuditReportQueryParams and AccessReportRow into audit.types.ts
  per module convention.
